requirements_test.txt
pytest
pytest-homeassistant-custom-component>=0.12.49
pytest-asyncio
pytest-cov
